# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RAG University Assistant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: documents
    description: Document management
  - name: chat
    description: Chat operations
  - name: analytics
    description: Analytics and stats
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all uploaded documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document for indexing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Document accepted and queued for indexing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  /documents/{id}/chunks:
    get:
      operationId: getDocumentChunks
      tags: [documents]
      summary: Get all chunks for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of chunks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentChunk"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  /documents/rebuild-index:
    post:
      operationId: rebuildIndex
      tags: [documents]
      summary: Rebuild the search index for all documents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Index rebuilt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RebuildResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  /chat/sessions:
    get:
      operationId: listChatSessions
      tags: [chat]
      summary: List all chat sessions
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create a new chat session
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  /chat/sessions/{sessionId}:
    delete:
      operationId: deleteChatSession
      tags: [chat]
      summary: Delete a chat session and its history
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  /chat/sessions/{sessionId}/messages:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat history for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  /chat/ask:
    post:
      operationId: askQuestion
      tags: [chat]
      summary: Ask a question (streaming SSE response)
      description: |
        Answers are streamed as Server-Sent Events, so this operation cannot be
        consumed by React Query; the SPA hand-writes a `fetch` reader for it.
        The frame contract is:

        ```
        data: {"sessionId": 12}                                    // first frame, only when a session was created
        data: {"content": "..."}                                   // repeated, one per token batch
        data: {"done":true,"messageId":34,"sources":[...],"confidence":0.82,"sessionId":12}
        data: {"error":"human-readable message"}                   // terminal on failure - ALWAYS sent
        :                                                          // heartbeat comment every 15s
        ```
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "200":
          description: Event stream of answer tokens, ending in a `done` or `error` frame
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  /chat/messages/{messageId}/feedback:
    post:
      operationId: submitFeedback
      tags: [chat]
      summary: Submit thumbs up/down feedback on a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "200":
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  /analytics/stats:
    get:
      operationId: getAnalyticsStats
      tags: [analytics]
      summary: Get overall analytics stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  /analytics/activity:
    get:
      operationId: getActivity
      tags: [analytics]
      summary: Get question volume per day for the last 14 days
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Daily question counts, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityPoint"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  /analytics/top-questions:
    get:
      operationId: getTopQuestions
      tags: [analytics]
      summary: Get most asked questions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Top questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopQuestion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  /analytics/suggested-questions:
    get:
      operationId: getSuggestedQuestions
      tags: [analytics]
      summary: Get suggested questions for students
      responses:
        "200":
          description: Suggested questions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Shared admin token from `ADMIN_TOKEN`

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid admin token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PayloadTooLarge:
      description: File exceeds the 20 MB upload limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string

    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Document:
      type: object
      required: [id, filename, originalName, mimeType, fileSize, chunkCount, createdAt, status]
      properties:
        id:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        chunkCount:
          type: integer
        status:
          type: string
          enum: [processing, ready, error]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string

    DocumentChunk:
      type: object
      required: [id, documentId, chunkText, chunkIndex, createdAt]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        chunkText:
          type: string
        pageNumber:
          type: ["integer", "null"]
        chunkIndex:
          type: integer
        createdAt:
          type: string

    RebuildResult:
      type: object
      required: [success, chunksIndexed]
      properties:
        success:
          type: boolean
        chunksIndexed:
          type: integer

    ChatSession:
      type: object
      required: [id, title, createdAt, messageCount]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
        messageCount:
          type: integer
        lastMessageAt:
          type: ["string", "null"]

    ChatMessage:
      type: object
      required: [id, sessionId, role, content, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        sources:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/Citation"
        confidence:
          type: ["number", "null"]
        feedback:
          type: ["string", "null"]
          enum: [helpful, not_helpful, null]
        createdAt:
          type: string

    Citation:
      type: object
      required: [documentName, pageNumber, chunkText]
      properties:
        documentName:
          type: string
        pageNumber:
          type: ["integer", "null"]
        chunkText:
          type: string
        score:
          type: ["number", "null"]

    QuestionInput:
      type: object
      required: [question]
      properties:
        question:
          type: string
          minLength: 1
        sessionId:
          type: ["integer", "null"]

    FeedbackInput:
      type: object
      required: [feedback]
      properties:
        feedback:
          type: string
          enum: [helpful, not_helpful]

    AnalyticsStats:
      type: object
      required:
        [
          totalDocuments,
          totalChunks,
          totalQuestions,
          totalSessions,
          helpfulRate,
          avgConfidence,
          documentsByStatus,
        ]
      properties:
        totalDocuments:
          type: integer
        totalChunks:
          type: integer
        totalQuestions:
          type: integer
        totalSessions:
          type: integer
        helpfulRate:
          type: number
        avgConfidence:
          type: number
        questionsToday:
          type: integer
        documentsByStatus:
          type: object
          required: [processing, ready, error]
          properties:
            processing:
              type: integer
            ready:
              type: integer
            error:
              type: integer

    ActivityPoint:
      type: object
      required: [date, questions]
      properties:
        date:
          type: string
        questions:
          type: integer

    TopQuestion:
      type: object
      required: [question, count]
      properties:
        question:
          type: string
        count:
          type: integer