# lib/api-zod/src/generated/types/analyticsStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RAG University Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsStatsDocumentsByStatus } from './analyticsStatsDocumentsByStatus';

export interface AnalyticsStats {
  totalDocuments: number;
  totalChunks: number;
  totalQuestions: number;
  totalSessions: number;
  helpfulRate: number;
  avgConfidence: number;
  questionsToday?: number;
  documentsByStatus: AnalyticsStatsDocumentsByStatus;
}