# lib/api-zod/src/generated/types/chatSession.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RAG University Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatSession {
  id: number;
  title: string;
  createdAt: string;
  messageCount: number;
  /** @nullable */
  lastMessageAt?: string | null;
}