# lib/api-zod/src/generated/types/citation.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RAG University Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Citation {
  documentName: string;
  /** @nullable */
  pageNumber: number | null;
  chunkText: string;
  /** @nullable */
  score?: number | null;
}