# lib/api-zod/src/generated/types/rebuildResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RAG University Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface RebuildResult {
  success: boolean;
  chunksIndexed: number;
}